refactor: rename retrieve tool to knowledge_search
"Retrieve" is ambiguous — doesn't tell the LLM (or the reader) what's
being retrieved or how. Renamed to knowledge_search to parallel the
existing web_search naming (domain_search) and to make the tool's
purpose obvious in prompt listings.

Full rename, all three levels:
  - Go package: plugins/tools/retrieve/ → plugins/tools/knowledge_search/
  - Plugin ID: nexus.tool.retrieve → nexus.tool.knowledge_search
  - Tool name (LLM-facing): retrieve → knowledge_search

Also updated the tool description to explicitly contrast it with
web_search, so the LLM picks knowledge_search for configured knowledge
and web_search for the open web.

Updates configs/rag.yaml (active list + config block + system prompt)
and CLAUDE.md (plugin listing + RAG section + profile summary) to match.
No on-disk state to migrate — this is pre-merge.
